Rename owned texture frame leases to handles
